move views actions in more files
